Mv功能完成一部分

+++ Lgs_module_login/src/main/java/com/example/lgs_module/bean/Account.kt
@@ -2,18 +2,18 @@ package com.example.lgs_module.bean
 
 data class Account(
     val anonimousUser: Boolean,
-    val ban: Int,
-    val baoyueVersion: Int,
+    val ban: Long,
+    val baoyueVersion: Long,
     val createTime: Long,
     val donateVersion: Int,
     val id: Long,
     val salt: String,
     val status: Int,
-    val tokenVersion: Int,
+    val tokenVersion: Long,
     val type: Int,
     val uninitialized: Boolean,
     val userName: String,
-    val vipType: Int,
+    val vipType: Long,
     val viptypeVersion: Long,
-    val whitelistAuthority: Int
+    val whitelistAuthority: Long
 )
