+++ Lgs_module_login/src/main/java/com/example/lgs_module/bean/Binding.kt
@@ -3,11 +3,11 @@ package com.example.lgs_module.bean
 data class Binding(
     val bindingTime: Long,
     val expired: Boolean,
-    val expiresIn: Int,
+    val expiresIn: Long,
     val id: Long,
-    val refreshTime: Int,
+    val refreshTime: Long,
     val tokenJsonStr: String,
-    val type: Int,
+    val type: Long,
     val url: String,
     val userId: Long
 )
