+++ Lgs_module_login/src/main/java/com/example/lgs_module/bean/Regitser.kt
@@ -3,7 +3,7 @@ package com.example.lgs_module.bean
 data class Regitser(
     val account: Account,
     val bindings: List<Binding>,
-    val code: Int,
+    val code: Long,
     val profile: Profile,
     val token: String
 )
